Added SlidingTabLayout Updated Register & Login screen

+++ finan/src/main/java/pierrydev/com/br/finan/domain/contracts/repositories/IEntryRepository.java
@@ -10,7 +10,7 @@ public interface IEntryRepository {
 
   Entry getById(int id);
 
-  List<Entry> getByLocalId(int localId);
+  List<Entry> getByLocalId(long localId);
 
   void create(Entry user);
